requirements.txt
pyyaml==6.0
pulumi>=3.0.0,<4.0.0
pulumi-gcp>=6.0.0,<7.0.0
jsonschema==4.17.3
confluent-kafka==2.1.0
requests==2.31.0
